Replace enum for spatial relationships with a std::bitset.
Note that the INTERSECTS relationship no longer gets its own bit,
since it is just the complement of the DISJOINT relationship. As a
result, all possible combinations of relationship bits are
meaningful, and relationship enumeration is more intuitive. For
example, the bitset for CONTAINS implies INTERSECTS (since the
bit for DISJOINT is not set).