Implement HTM / region intersection.
It is now possible to compute the indexes of the trixels
intersecting a given region, optionally using an adaptive
subdivision level to limit computational costs. This is
exposed by making the HTM code an implementation of the
new Pixelization interface.

The HTM specific documentation is moved to a markdown file,
and .md files are flagged for doxygen processing. This also
allows the top-level github README to be reused as the
doxygen mainpage.