Adopt PEP 517 for build dependencies
This adds a pyproject.toml file. This lets us specify Python
dependencies that are needed to build the package (before setup.py is
run). This replaces the previous approach of using "setup_requires"
dependencies which are only installed *after* the setup function is run,
which is too late in this case (setup_requires is intended more for
setuptools plugins that integrate with a setuptools hook).

Since setup.py is integral here, I've also moved all packaging-related
content out of setup.cfg to setup.py to simplify matters. This may not
be necessary, and its possible that we can have packaging config in both
the setup.cfg and setup.py, but I wanted to avoid this opportunity for
confusion for now.