Use tox in GitHub Actions build_and_test
This does two things. Firstly, it adopts tox for running the build and
test, which is beneficial because it makes CI similar to the development
testing set up.

Second, it removes the artificial crutch of the requirements.txt, which
can't be used if this package is expected to be installed by downstream
packages using pip alone.