Adopt pyproject.toml for setting build reqs
pyroject.toml and PEP 517 is really the only way to ensure that certain
python dependencies are present *before* a package is built.

This replaces the setup_requires section in setup.cfg (it didn't install
build dependencies in time) and requirements.txt (it's not a viable
solution for installing dependencies for a library).

Update GitHub Actions to use this installation method.