Special-case Circle comparisons that involve equal circles.
It's understandable that the general implementation of these operations
doesn't handle this case the way we'd like due to round-off error, and
the old behavior isn't even wrong, given their clearly-documented
limited guarantees.  But it's very nice for unit tests at least to
satisfy "equality is equivalent to CONTAINS and WITHIN".